Allow decimal flat rates and read harvest count messages
- flatRate is now a number rather than an integer (schema change agreed
  on 2026-09-17), because Scorpia's offspring is 1/2,015.75. Rates show
  as written: whole numbers grouped, decimals kept. baseChance and
  contributionRange stay integers.
- Chat counts follow RuneLite 1.12.39's KILLCOUNT_PATTERN more closely:
  boss names wrapped in colour tags are unwrapped, and 'harvest count'
  messages record '<activity>_harvests' (Herbiboar has no collection
  log counter, so its chat message is the only source).
- Temporary debug logging of chompy chat messages, to verify the ogre
  bow kill check format before relying on it.

+++ src/main/java/com/pethunter/data/PetRepository.java
@@ -257,7 +257,7 @@ public final class PetRepository
 				.label(optString(obj, "label"))
 				.rateModel(rateModel)
 				.baseChance(optInteger(obj, "baseChance"))
-				.flatRate(optInteger(obj, "flatRate"))
+				.flatRate(optDouble(obj, "flatRate"))
 				.contributionRange(parseContributionRange(obj))
 				.minLevel(optInteger(obj, "minLevel"))
 				.counterKey(optString(obj, "counterKey"))
